workdir/pyode/requirements.txt
[pip]
Py3ODE >= 1.2.0.dev15
pygame >= 2.0.0
PyOpenGL >= 3.1.5
PyOpenGL_accelerate >= 3.1.5
vpython >= 7.6.1

[apt-get]
libglu1-mesa-dev freeglut3-dev mesa-common-dev